Completed implementation for this User Story

+++ src/main/java/ca/gc/tri_agency/granting_data/service/FundingCycleService.java
@@ -43,6 +43,6 @@ public interface FundingCycleService {
 
 	FundingCycleProjection findFundingCycleForConfirmDeleteFC(Long fcId) throws AccessDeniedException;
 
-	List<FundingCycleProjection> findFundingCyclesForCalendar(int plusMinusMonth);
+	List<FundingCycleProjection> findFundingCyclesForCalendar(long plusMinusMonth);
 
 }
